requirements.txt
fastapi
uvicorn
python-dotenv
openai
requests
tavily-python     

faiss-cpu
sentence-transformers
numpy
pypdf
python-multipart